+++ hosts/darwin/common-darwin.nix
@@ -6,6 +6,7 @@
   lib,
   pkgs,
   nvf,
+  qmd,
   nix-secrets,
   ...
 }:
@@ -126,6 +127,7 @@ in
     backupCommand = "${pkgs.trash-cli}/bin/trash";
     sharedModules = [
       nvf.homeManagerModules.default
+      qmd.homeModules.default
     ];
     extraSpecialArgs = { inherit nix-secrets; };
     users.${primaryUser} = {